split out shared api fns
